--- kancolle/orm.py
@@ -1,316 +1,0 @@
-from . import db
-from . import UserMixin, RoleMixin
-
-role__user = db.Table('role__user',
-                      db.Column('user_id', db.Integer, db.ForeignKey('user.id')),
-                      db.Column('role_id', db.Integer, db.ForeignKey('role.id'))
-                      )
-
-class KCModel():
-    def byId(self,id=0):
-        return db.session.query(type(self)).get(id)
-
-
-class RecipeMap(db.Model,KCModel):
-    __tablename__ = 'recipe'
-    id = db.Column(db.Integer, primary_key=True)
-    ship = db.relationship("Ship", uselist=False)
-    ship_id = db.Column(db.Integer, db.ForeignKey('ship.id'))
-
-    chance = db.Column(db.Integer, default=1)
-
-    minfuel = db.Column(db.Integer, nullable=False, default=30)
-    maxfuel = db.Column(db.Integer, nullable=False, default=30)
-    minammo = db.Column(db.Integer, nullable=False, default=30)
-    maxammo = db.Column(db.Integer, nullable=False, default=30)
-    minsteel = db.Column(db.Integer, nullable=False, default=30)
-    maxsteel = db.Column(db.Integer, nullable=False, default=30)
-    minbaux = db.Column(db.Integer, nullable=False, default=30)
-    maxbaux = db.Column(db.Integer, nullable=False, default=30)
-
-
-class FleetMap(db.Model,KCModel):
-    __tablename__ = 'fleet'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String, default="Unnamed")
-    admiral_id = db.Column(db.Integer, db.ForeignKey('admiral.id'))
-
-    ships = db.relationship("AdmiralShip", lazy='dynamic', order_by='AdmiralShip.local_fleet_num')
-
-
-class DockMap(db.Model,KCModel):
-    __tablename__ = 'dock'
-    id = db.Column(db.Integer, primary_key=True)
-    state = db.Column(db.Integer)
-    complete = db.Column(db.BigInteger, nullable=True)
-    fuel = db.Column(db.Integer, nullable=True)
-    ammo = db.Column(db.Integer, nullable=True)
-    steel = db.Column(db.Integer, nullable=True)
-    baux = db.Column(db.Integer, nullable=True)
-    cmats = db.Column(db.Integer, nullable=True)
-
-    ship_id = db.Column(db.Integer, db.ForeignKey("admiral_ship.id"))
-    ship = db.relationship("AdmiralShip", uselist=False)
-
-    admiral_id = db.Column(db.Integer, db.ForeignKey("admiral.id"))
-
-
-class AdmiralShipMap(db.Model,KCModel):
-    __tablename__ = 'admiral_ship'
-    id = db.Column(db.Integer, primary_key=True)
-    admiral_id = db.Column(db.Integer, db.ForeignKey('admiral.id'))
-    ship = db.relationship("Ship", uselist=False)
-    ship_id = db.Column(db.Integer, db.ForeignKey('ship.id'))
-
-    fleet_id = db.Column(db.Integer, db.ForeignKey('fleet.id'))
-
-    local_fleet_num = db.Column(db.Integer)
-    local_ship_num = db.Column(db.Integer, nullable=False)
-
-    # Unique ship-specific attributes
-    ammo = db.Column(db.Integer)
-    fuel = db.Column(db.Integer)
-    fatigue = db.Column(db.Integer, default=49)
-
-    exp = db.Column(db.Integer)
-    level = db.Column(db.Integer)
-
-    repair_base = db.Column(db.String())
-
-    current_hp = db.Column(db.Integer)  # Oh dear.
-
-    # Ship stats
-    luck = db.Column(db.Integer, default=0)
-    luck_eq = db.Column(db.Integer, default=0)
-    firepower = db.Column(db.Integer, default=0)
-    firepower_eq = db.Column(db.Integer, default=0)
-    armour = db.Column(db.Integer, default=0)
-    torpedo = db.Column(db.Integer, default=0)
-    torpedo_eq = db.Column(db.Integer, default=0)
-    antiair = db.Column(db.Integer, default=0)
-    antiair_eq = db.Column(db.Integer, default=0)
-    antisub = db.Column(db.Integer, default=0)
-    evasion = db.Column(db.Integer, default=0)
-
-    heartlocked = db.Column(db.Boolean, default=False)
-
-    active = db.Column(db.Boolean, default=False, nullable=False)
-
-    items = db.relationship("Item", backref="admiral_ship", lazy="dynamic")
-
-    def __str__(self):
-        return self.ship.name
-
-class ShipMap(db.Model,KCModel):
-    __tablename__ = 'ship'
-    id = db.Column(db.Integer, primary_key=True)
-    afterlv = db.Column(db.Integer, nullable=True)
-    aftership_num = db.Column(db.Integer, nullable=True)
-    remodel_cost = db.Column(db.String())
-
-    rarity = db.Column(db.Integer)
-    broken = db.Column(db.String())
-
-    ammo_max = db.Column(db.Integer)
-    fuel_max = db.Column(db.Integer)
-
-    name = db.Column(db.String)
-    number = db.Column(db.Integer)
-    stype = db.Column(db.Integer)
-
-    voicef = db.Column(db.Integer)
-
-    itemslots = db.Column(db.Integer)
-
-    modern_use = db.Column(db.String)
-    # Minimums
-    luck_base = db.Column(db.Integer, default=0)
-    firepower_base = db.Column(db.Integer, default=0)
-    torpedo_base = db.Column(db.Integer, default=0)
-    armour_base = db.Column(db.Integer, default=0)
-    antiair_base = db.Column(db.Integer, default=0)
-    antisub_base = db.Column(db.Integer, default=0)
-    los_base = db.Column(db.Integer, default=0)
-    evasion_base = db.Column(db.Integer, default=0)
-    hp_base = db.Column(db.Integer, default=0)
-
-    # Maximums
-    luck_max = db.Column(db.Integer, default=0)
-    firepower_max = db.Column(db.Integer, default=0)
-    torpedo_max = db.Column(db.Integer, default=0)
-    armour_max = db.Column(db.Integer, default=0)
-    antiair_max = db.Column(db.Integer, default=0)
-    antisub_max = db.Column(db.Integer, default=0)
-    maxslots = db.Column(db.Integer, default=0)
-    maxlos = db.Column(db.Integer, default=0)
-    evasion_max = db.Column(db.Integer, default=0)
-
-    maxhp = db.Column(db.Integer)
-    srange = db.Column(db.Integer, default=0)
-
-    kai = db.Column(db.Boolean, default=False)
-
-    # Messages
-    getmsg = db.Column(db.Text())
-    buildtime = db.Column(db.Integer)
-
-    maxplanes = db.Column(db.Integer())
-
-    def __str__(self):
-        return self.name
-
-class AdmiralQuestMap(db.Model,KCModel):
-    """
-    progress: None,50,80
-    state: Inactive, Active, Complete
-    """
-    __tablename__ = 'admiral_quest'
-    admiral_id = db.Column(db.Integer, db.ForeignKey('admiral.id'),primary_key=True)
-    quest_id = db.Column(db.Integer, db.ForeignKey('quest.id'),primary_key=True)
-
-    progress = db.Column(db.Integer)
-    state = db.Column(db.Integer)
-
-class AdmiralSortieMap(db.Model,KCModel):
-    __tablename__ = 'admiral_sortie'
-    admiral_id = db.Column(db.Integer, db.ForeignKey('admiral.id'),primary_key=True)
-    sortie_id = db.Column(db.Integer, db.ForeignKey('sortie.id'),primary_key=True)
-    cleared = db.Column(db.Integer,default=0)
-    defeat_count = db.Column(db.Integer)   
-
-class AdmiralMap(db.Model,KCModel):
-    __tablename__ = 'admiral'
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
-    last_login = db.Column(db.DateTime)
-
-    level = db.Column(db.Integer, default=1)
-    rank = db.Column(db.Integer, default=8)
-    experience = db.Column(db.Integer, default=0)
-    tutorial_progress = db.Column(db.Integer, default=0)
-
-    furniture = db.Column(db.String(), default="1,1,1,1,1,1")
-    furniture_coins = db.Column(db.Integer, default=0)
-
-    max_ships = db.Column(db.Integer, default=5000)
-    max_equips = db.Column(db.Integer, default=10000)
-    max_furniture = db.Column(db.Integer, default=10000)
-    available_fleets = db.Column(db.Integer, default=1)
-    available_cdocks = db.Column(db.Integer, default=4)
-    available_rdocks = db.Column(db.Integer, default=4)
-
-    sortie_successes = db.Column(db.Integer, default=0)
-    sortie_total = db.Column(db.Integer, default=0)
-    expedition_successes = db.Column(db.Integer, default=0)
-    expedition_total = db.Column(db.Integer, default=0)
-    pvp_successes = db.Column(db.Integer, default=0)
-    pvp_total = db.Column(db.Integer, default=0)
-
-    admiral_ships = db.relationship("AdmiralShip", backref='admiral', lazy='dynamic')
-
-    resources = db.Column(db.String())
-
-    fleets = db.relationship("Fleet", backref='admiral', lazy='dynamic')
-
-    sorties = db.relationship("AdmiralSortie", backref='admiral', lazy='dynamic')
-
-    quests = db.relationship("AdmiralQuest", backref='admiral', lazy='dynamic')
-
-    # If this is false...
-    # 1) api_req_init is enabled
-    #   2) setup() is enabled, which means a new fleet will be automatically added, and two docks of each kind
-    # If this is true...
-    #   1) api_req_init will be disabled
-    #   2) setup() is disabled.
-    setup = db.Column(db.Boolean())
-    docks = db.relationship("Dock", backref='admiral', lazy='dynamic', order_by='Dock.id')
-
-    lastaction = db.Column(db.DateTime)
-
-    def __str__(self):
-        return "Admiral " + self.user.nickname
-
-
-class ItemMap(db.Model,KCModel):
-    __tablename__ = 'item'
-    id = db.Column(db.Integer, primary_key=True)
-
-    admiral_ship_id = db.Column(db.Integer, db.ForeignKey("admiral_ship.id"))
-
-    name = db.Column(db.String(255), primary_key=True)
-    desc = db.Column(db.Text)
-
-    usable = db.Column(db.Boolean, default=False)
-
-    increase_fp = db.Column(db.Boolean)
-    increase_aa = db.Column(db.Boolean)
-    increase_tp = db.Column(db.Boolean)
-    increase_ar = db.Column(db.Boolean)
-
-    increase_amount = db.Column(db.Integer, nullable=False)
-
-    rarity = db.Column(db.Integer, nullable=False)
-
-    sortno = db.Column(db.Integer)
-
-
-class UserMap(db.Model,KCModel, UserMixin):
-    __tablename__ = 'user'
-    id = db.Column(db.Integer, primary_key=True)
-    active = db.Column(db.Boolean)
-
-    email = db.Column(db.String(255), unique=True, nullable=False)
-    password = db.Column(db.String(255), nullable=False)
-    confirmed_at = db.Column(db.DateTime())
-
-    nickname = db.Column(db.String(255), unique=True, nullable=False)
-    api_token = db.Column(db.String(40), default=lambda: util.generate_api_token(), unique=True)
-
-    admiral = db.relationship("Admiral", backref='user', uselist=False)
-
-    roles = db.relationship('Role', secondary=role__user, backref=db.backref('users', lazy='dynamic'))  # Fix roles
-
-    def __str__(self):
-        return self.email
-
-
-
-class RoleMap(db.Model,KCModel, RoleMixin):
-    __tablename__ = 'role'
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), unique=True, nullable=False)
-    description = db.Column(db.Text)
-
-    def __str__(self):
-        return self.name
-
-class SortieMap(db.Model,KCModel):
-    __tablename__ = 'sortie'
-    id = db.Column(db.Integer, primary_key=True)
-
-    is_boss = db.Column(db.Integer, default=0)
-    level = db.Column(db.String(10))
-
-class QuestMap(db.Model,KCModel):
-    __tablename__ = 'quest'
-    """
-    no: quest number
-    frequency: Once,Daily,Weekly
-    category: Compostition, Exercise, etc.
-    bonus_flag: ?
-    invalid_flag: ?
-    get_material: "[val,val,val,val]"
-    code: A1,B20, etc. Not required. Maybe easier to maintain?
-    """
-    id = db.Column(db.Integer,primary_key=True)
-
-    no = db.Column(db.Integer)
-    category = db.Column(db.Integer)
-    frequency = db.Column(db.Integer) #Kancolle calls it 'type'
-    title = db.Column(db.String(255))
-    detail = db.Column(db.Text)
-    bonus_flag = db.Column(db.Integer)
-    invalid_flag = db.Column(db.Integer)
-    get_material = db.Column(db.String(25))
-    code = db.Column(db.String(3))
-
